Cover quiet biomarker refresh states
