+++ apps/web/app/design/browser-vault-loading-transitions-study.tsx
@@ -2,9 +2,16 @@
 
 import type {
   BrowserVaultDeviceMetricSummary,
+  BrowserVaultLabBiomarkerDetail,
   BrowserVaultMeasuredBiomarker,
+  BrowserVaultPresentedLabResultRow,
 } from "@murphai/query/browser-biomarkers";
 
+import {
+  BiomarkerDetailContent,
+  BiomarkerDetailShell,
+  EmptyBiomarkerDetailCard,
+} from "@/src/components/biomarkers/lab-biomarker-detail-view";
 import {
   BiomarkerListSkeleton,
   DeviceMetricsSection,
@@ -155,6 +162,43 @@ const DESIGN_LAB_GROUP: MeasuredBiomarkerGroup = {
   label: "Nutrients & fatty acids",
 };
 
+const DESIGN_LAB_ROWS: readonly BrowserVaultPresentedLabResultRow[] = [
+  {
+    ...DESIGN_LAB_BIOMARKER.latest,
+    date: "2025-11-01",
+    id: "design-ferritin-result-2025",
+    normalizedValue: 58,
+    observedAt: "2025-11-01T08:00:00.000Z",
+    value: 58,
+  },
+  {
+    ...DESIGN_LAB_BIOMARKER.latest,
+    date: "2026-05-01",
+    id: "design-ferritin-result-2026-05",
+    normalizedValue: 68,
+    observedAt: "2026-05-01T08:00:00.000Z",
+    value: 68,
+  },
+  DESIGN_LAB_BIOMARKER.latest,
+];
+
+const DESIGN_LAB_DETAIL: BrowserVaultLabBiomarkerDetail = {
+  biomarkerKey: "design-ferritin",
+  chartSeries: DESIGN_LAB_ROWS.map((row) => ({
+    date: row.date,
+    observedAt: row.observedAt,
+    rowId: row.id,
+    unit: "ng/mL",
+    value: row.normalizedValue ?? row.value ?? 0,
+  })),
+  comparableUnit: "ng/mL",
+  displayName: "Ferritin",
+  hasIncompatibleHistory: false,
+  latest: DESIGN_LAB_BIOMARKER.latest,
+  metricKey: "design-ferritin",
+  rows: [...DESIGN_LAB_ROWS],
+};
+
 const noopRetry = async () => undefined;
 
 function StudyState({
@@ -303,12 +347,39 @@ export function BrowserVaultLoadingTransitionsStudy() {
       </TransitionStudy>
 
       <TransitionStudy
-        description="The private result route reserves the latest reading, chart, and year-grouped ledger so the page keeps its shape while saved labs arrive."
+        description="The private result route reserves the latest reading, chart, and year-grouped ledger while loading, then keeps usable saved results or the empty state quiet during automatic refresh."
         title="Biomarker result detail"
         transition="biomarker-result-detail"
       >
         <StudyState label="Loading" state="loading">
           <BiomarkerDetailSkeleton />
+        </StudyState>
+        <StudyState label="Stale saved result" state="stale-populated">
+          <BiomarkerDetailShell
+            chatAction={null}
+            detail={DESIGN_LAB_DETAIL}
+            summary="Synthetic iron-storage marker history for layout review."
+          >
+            <BiomarkerDetailContent detail={DESIGN_LAB_DETAIL} fallbackRanges={[]} />
+          </BiomarkerDetailShell>
+        </StudyState>
+        <StudyState label="Refreshing saved result" state="refresh-pending-populated">
+          <BiomarkerDetailShell
+            chatAction={null}
+            detail={DESIGN_LAB_DETAIL}
+            summary="Synthetic iron-storage marker history for layout review."
+          >
+            <BiomarkerDetailContent detail={DESIGN_LAB_DETAIL} fallbackRanges={[]} />
+          </BiomarkerDetailShell>
+        </StudyState>
+        <StudyState label="Stale empty result" state="stale-empty">
+          <BiomarkerDetailShell chatAction={null} detail={null} summary={null}>
+            <EmptyBiomarkerDetailCard
+              authRequired={false}
+              preparing={false}
+              uploadLabsAction={null}
+            />
+          </BiomarkerDetailShell>
         </StudyState>
       </TransitionStudy>
     </div>
